refactor(bedrock): ♻️ rename bedrock-converse to amazon-bedrock
