+++ src/amazon-bedrock/bedrock-chat-language-model.ts
