+++ src/amazon-bedrock/bedrock-chat-settings.ts
