+++ src/amazon-bedrock/bedrock-error.ts
