+++ src/amazon-bedrock/bedrock-event-stream-response-handler.test.ts
