+++ src/amazon-bedrock/bedrock-event-stream-response-handler.ts
