+++ src/amazon-bedrock/bedrock-prepare-tools.ts
