+++ src/amazon-bedrock/convert-to-bedrock-chat-messages.ts
