+++ src/amazon-bedrock/map-bedrock-finish-reason.ts
